+++ library/src/blas2/rocblas_trsv_kernels.cpp
@@ -790,6 +790,31 @@ rocblas_trsv_device(rocblas_int    n,
 #endif
 }
 
+// this file uses kernels some defined above so inlined here
+#include "rocblas_trsv_big_batch_kernel.hpp"
+
+/**
+ * @brief Heuristic to determine if big batch kernel should be used
+ * 
+ * The big batch kernel is beneficial when:
+ * - batch_count is much larger than N
+ * - N is not big
+ * 
+ * @param n Matrix dimension
+ * @param batch_count Number of batched operations
+ * @return true if big batch kernel should be used
+ */
+inline bool should_use_big_batch_kernel(rocblas_int n, rocblas_int batch_count)
+{
+    // Threshold: use big batch kernel when batch_count > 16*n
+    constexpr rocblas_int BIG_BATCH_TO_N_RATIO = 16;
+
+    // Don't use big batch kernel for large matrices (matrix split creates sufficient work)
+    constexpr rocblas_int MAX_N_FOR_BIG_BATCH = 128;
+
+    return (n < MAX_N_FOR_BIG_BATCH) && (batch_count > BIG_BATCH_TO_N_RATIO * n);
+}
+
 template <rocblas_int DIM_X, typename T, typename TConstPtr, typename TPtr>
 rocblas_status rocblas_internal_trsv_substitution_template(rocblas_handle    handle,
                                                            rocblas_fill      uplo,
@@ -810,6 +835,44 @@ rocblas_status rocblas_internal_trsv_substitution_template(rocblas_handle    han
 {
     if(!n || !batch_count)
         return rocblas_status_success;
+
+    // Check if we should use the z-batched kernel
+    if(should_use_big_batch_kernel(n, batch_count))
+    {
+        // TODO: Select block size based on data type requires further tuning
+        constexpr rocblas_int SDCTRSV_BB_NB = 4;
+        constexpr rocblas_int ZTRSV_BB_NB   = 4;
+
+#define TRSV_BATCHED_TEMPLATE_PARAMS                                                       \
+    handle, uplo, transA, diag, n, dA, offset_A, lda, stride_A, alpha, dx, offset_x, incx, \
+        stride_x, batch_count, w_completed_sec
+
+        // Use z-dimension batched kernel for high batch count scenarios
+        if constexpr(std::is_same_v<T, float>)
+        {
+            return rocblas_internal_trsv_substitution_big_batch_template<SDCTRSV_BB_NB, T>(
+                TRSV_BATCHED_TEMPLATE_PARAMS);
+        }
+        else if constexpr(std::is_same_v<T, double>)
+        {
+            return rocblas_internal_trsv_substitution_big_batch_template<SDCTRSV_BB_NB, T>(
+                TRSV_BATCHED_TEMPLATE_PARAMS);
+        }
+        else if constexpr(std::is_same_v<T, rocblas_float_complex>)
+        {
+            return rocblas_internal_trsv_substitution_big_batch_template<SDCTRSV_BB_NB, T>(
+                TRSV_BATCHED_TEMPLATE_PARAMS);
+        }
+        else if constexpr(std::is_same_v<T, rocblas_double_complex>)
+        {
+            return rocblas_internal_trsv_substitution_big_batch_template<ZTRSV_BB_NB, T>(
+                TRSV_BATCHED_TEMPLATE_PARAMS);
+        }
+
+#undef TRSV_BATCHED_TEMPLATE_PARAMS
+
+        return rocblas_status_internal_error;
+    }
 
     offset_x = incx < 0 ? offset_x + incx * (1 - n) : offset_x;
 
@@ -983,14 +1046,12 @@ rocblas_status rocblas_internal_trsv_substitution_template(rocblas_handle    han
             }
         }
     }
+
 #undef TRSV_TEMPLATE_PARAMS
 
     return rocblas_status_success;
 }
 
-#define TRSV_TEMPLATE_PARAMS                                                                 \
-    handle, uplo, transA, diag, n, dA, offset_A, lda, stride_A, nullptr, dx, offset_x, incx, \
-        stride_x, batch_count, w_completed_sec
 template <typename T>
 ROCBLAS_INTERNAL_EXPORT_NOINLINE rocblas_status
     rocblas_internal_trsv_template(rocblas_handle    handle,
@@ -1009,6 +1070,10 @@ ROCBLAS_INTERNAL_EXPORT_NOINLINE rocblas_status
                                    rocblas_int       batch_count,
                                    rocblas_int*      w_completed_sec)
 {
+#define TRSV_TEMPLATE_PARAMS                                                                 \
+    handle, uplo, transA, diag, n, dA, offset_A, lda, stride_A, nullptr, dx, offset_x, incx, \
+        stride_x, batch_count, w_completed_sec
+
     if constexpr(std::is_same_v<T, float>)
         return rocblas_internal_trsv_substitution_template<ROCBLAS_SDCTRSV_NB, T>(
             TRSV_TEMPLATE_PARAMS);
@@ -1023,6 +1088,8 @@ ROCBLAS_INTERNAL_EXPORT_NOINLINE rocblas_status
             TRSV_TEMPLATE_PARAMS);
 
     return rocblas_status_internal_error;
+
+#undef TRSV_TEMPLATE_PARAMS
 }
 
 template <typename T>
@@ -1043,23 +1110,27 @@ ROCBLAS_INTERNAL_EXPORT_NOINLINE rocblas_status
                                            rocblas_int       batch_count,
                                            rocblas_int*      w_completed_sec)
 {
+#define TRSV_BATCHED_TEMPLATE_PARAMS                                                         \
+    handle, uplo, transA, diag, n, dA, offset_A, lda, stride_A, nullptr, dx, offset_x, incx, \
+        stride_x, batch_count, w_completed_sec
+
     if constexpr(std::is_same_v<T, float>)
         return rocblas_internal_trsv_substitution_template<ROCBLAS_SDCTRSV_NB, T>(
-            TRSV_TEMPLATE_PARAMS);
+            TRSV_BATCHED_TEMPLATE_PARAMS);
     else if constexpr(std::is_same_v<T, double>)
         return rocblas_internal_trsv_substitution_template<ROCBLAS_SDCTRSV_NB, T>(
-            TRSV_TEMPLATE_PARAMS);
+            TRSV_BATCHED_TEMPLATE_PARAMS);
     else if constexpr(std::is_same_v<T, rocblas_float_complex>)
         return rocblas_internal_trsv_substitution_template<ROCBLAS_SDCTRSV_NB, T>(
-            TRSV_TEMPLATE_PARAMS);
+            TRSV_BATCHED_TEMPLATE_PARAMS);
     else if constexpr(std::is_same_v<T, rocblas_double_complex>)
         return rocblas_internal_trsv_substitution_template<ROCBLAS_ZTRSV_NB, T>(
-            TRSV_TEMPLATE_PARAMS);
+            TRSV_BATCHED_TEMPLATE_PARAMS);
+
+#undef TRSV_BATCHED_TEMPLATE_PARAMS
 
     return rocblas_status_internal_error;
 }
-
-#undef TRSV_TEMPLATE_PARAMS
 
 template <typename T, typename U>
 rocblas_status rocblas_internal_trsv_check_numerics(const char*       function_name,